llama-index-integrations/vector_stores/llama-index-vector-stores-valkey/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "docker>=7.1.0,<8",
    "ipython==8.10.0",
    "jupyter>=1.0.0,<2",
    "mypy==0.991",
    "pre-commit==3.2.0",
    "pylint==2.15.10",
    "pytest==7.2.1",
    "pytest-asyncio>=0.21.0",
    "pytest-mock==3.11.1",
    "ruff==0.11.11",
    "types-Deprecated>=0.1.0",
    "types-PyYAML>=6.0.12.12,<7",
    "types-protobuf>=4.24.0.4,<5",
    "types-requests==2.28.11.8",
    "types-setuptools==67.1.0.0",
    "black[jupyter]<=23.9.1,>=23.7.0",
    "codespell[toml]>=v2.2.6",
    "diff-cover>=9.2.0",
    "pytest-cov>=6.1.1",
    "hypothesis>=6.0.0",
]

[project]
name = "llama-index-vector-stores-valkey"
version = "0.1.0"
description = "llama-index vector_stores valkey integration"
authors = [{name = "Daria Korenieva", email = "daric2612@gmail.com"}]
requires-python = ">=3.9,<3.14"
readme = "README.md"
license = "MIT"
dependencies = [
    "valkey-glide>=2.2.7",
    "valkey-glide-sync>=2.1.0",
    "llama-index-core>=0.13.0,<0.15",
    "llama-index-embeddings-openai>=0.5.1",
]

[tool.codespell]
check-filenames = true
check-hidden = true
skip = "*.csv,*.html,*.json,*.jsonl,*.pdf,*.txt,*.ipynb"

[tool.hatch.build.targets.sdist]
include = ["llama_index/"]
exclude = ["**/BUILD"]

[tool.hatch.build.targets.wheel]
include = ["llama_index/"]
exclude = ["**/BUILD"]

[tool.llamahub]
contains_example = false
import_path = "llama_index.vector_stores.valkey"

[tool.llamahub.class_authors]
ValkeyVectorStore = "llamaindex"

[tool.mypy]
disallow_untyped_defs = true
exclude = ["_static", "build", "examples", "notebooks", "venv"]
ignore_missing_imports = true
python_version = "3.9"

[tool.uv.sources]
llama-index-core = {path = "../../../llama-index-core", editable = true}
